refactor(cli): abstract file funcs to package
Use internal/cache and internal/file instead of lwint

ALLY-844
