+++ cli/cmd/migration.go
@@ -28,7 +28,8 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/viper"
 
-	lwint "github.com/lacework/go-sdk/internal/lacework"
+	"github.com/lacework/go-sdk/internal/cache"
+	"github.com/lacework/go-sdk/internal/file"
 	"github.com/lacework/go-sdk/lwconfig"
 )
 
@@ -120,7 +121,7 @@ func (c *cliState) Migrations() (err error) {
 	// if the configuration file does not exist, most likely the user
 	// is executing the CLI via env variables or flags, update feature
 	// field and exit migration
-	if !lwint.FileExists(viper.ConfigFileUsed()) {
+	if !file.FileExists(viper.ConfigFileUsed()) {
 		c.Log.Debugw("config file not found, skipping profile migration")
 		c.Event.AddFeatureField("config_file", "not_found")
 		return nil
@@ -159,7 +160,7 @@ func createConfigurationBackup() (string, error) {
 		return "", err
 	}
 
-	cacheDir, err := lwint.CacheDir()
+	cacheDir, err := cache.CacheDir()
 	if err != nil {
 		return "", err
 	}
